+++ src/component/AppLayout/Sider/ProjectExplorer.jsx
@@ -18,6 +18,7 @@ export class ProjectExplorer extends React.Component {
       removeSettingsProject: PropTypes.func.isRequired,
       saveSuite: PropTypes.func.isRequired,
       setSuite: PropTypes.func.isRequired,
+      setApp: PropTypes.func.isRequired,
       removeAppTab: PropTypes.func.isRequired,
       resetSuite: PropTypes.func.isRequired,
       resetProject: PropTypes.func.isRequired
@@ -63,6 +64,12 @@ export class ProjectExplorer extends React.Component {
 
     if ( projectDirectory === dir ) {
       const menu = new Menu();
+
+      menu.append( new MenuItem({
+        label: "New Suite...",
+        click: this.onNewSuite
+      }) );
+
       menu.append( new MenuItem({
         label: "Delete from the list",
         click: async () => {
@@ -116,7 +123,6 @@ export class ProjectExplorer extends React.Component {
       }
     }) );
 
-
     menu.append( new MenuItem({
       label: "Delete from the list",
       click: async () => {
@@ -151,6 +157,41 @@ export class ProjectExplorer extends React.Component {
   }
 
 
+  onNewSuite = async () => {
+    const { projectDirectory } = this.props;
+    if ( !projectDirectory ) {
+      return;
+    }
+    if ( this.props.suiteModified ) {
+      await confirmUnsavedChanges({
+        saveSuite: this.props.action.saveSuite,
+        setSuite: this.props.action.setSuite
+      });
+    }
+    this.props.action.setApp({ newSuiteModal: true });
+  }
+
+  onNewProject = async () => {
+    if ( this.props.suiteModified ) {
+      await confirmUnsavedChanges({
+        saveSuite: this.props.action.saveSuite,
+        setSuite: this.props.action.setSuite
+      });
+    }
+    this.props.action.setApp({ newProjectModal: true });
+  }
+
+  onOpenProject = async () => {
+    if ( this.props.suiteModified ) {
+      await confirmUnsavedChanges({
+        saveSuite: this.props.action.saveSuite,
+        setSuite: this.props.action.setSuite
+      });
+    }
+    this.props.action.setApp({ openProjectModal: true });
+  }
+
+
   render() {
     const { projectDirectory  } = this.props,
           projects = Object.entries( this.props.projects )
@@ -163,7 +204,19 @@ export class ProjectExplorer extends React.Component {
       <ErrorBoundary>
         <div id="cProjectNavigator" className="project-navigator">
           <h2>Projects
-            <span className="project-navigator__add"></span>
+
+            <a
+              tabIndex={1} role="button"
+              title="Create a project"
+              className="project-navigator__open" onClick={ this.onNewProject }>
+              <Icon type="folder-add" theme="filled" /></a>
+
+            <a
+              tabIndex={2} role="button"
+              title="Add a project into the list"
+              className="project-navigator__add" onClick={ this.onOpenProject }>
+              <Icon type="folder-open" theme="filled" /></a>
+
           </h2>
 
           <nav className="project-navigator__nav">